+++ src/util/string_utils.cpp
@@ -33,7 +33,8 @@ std::string strip_string(const std::string &s)
     =std::find_if_not(s.rbegin(), s.rend(), pred);
   std::string::size_type j=std::distance(right, s.rend())-1;
 
-  return s.substr(i, (j-i+1));
+  return s.substr(
+    static_cast<std::size_t>(i), static_cast<std::size_t>(j - i + 1));
 }
 
 /// Given a string s, split into a sequence of substrings when separated by
